fix(sidekick/rust): no repo metadata for showcase
The showcase client does not need a `.repo-metadata.json` file. The file is not
used, and generates spurious warnings.
