+++ internal/librarian/rust/generate.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/googleapis/librarian/internal/command"
 	"github.com/googleapis/librarian/internal/config"
-	"github.com/googleapis/librarian/internal/repometadata"
 	"github.com/googleapis/librarian/internal/serviceconfig"
 	"github.com/googleapis/librarian/internal/sidekick/parser"
 	sidekickrust "github.com/googleapis/librarian/internal/sidekick/rust"
@@ -83,7 +82,7 @@ func Generate(ctx context.Context, cfg *config.Config, library *config.Library, 
 	if err := sidekickrust.Generate(ctx, model, library.Output, modelConfig); err != nil {
 		return err
 	}
-	if len(model.Services) > 0 {
+	if needsRepoMetadata(model, library) {
 		repoMetadata, err := createRepoMetadata(cfg, library, sources)
 		if err != nil {
 			return err
@@ -96,23 +95,6 @@ func Generate(ctx context.Context, cfg *config.Config, library *config.Library, 
 		validate(ctx, library.Output)
 	}
 	return nil
-}
-
-func createRepoMetadata(cfg *config.Config, library *config.Library, sources *sources.Sources) (*repometadata.RepoMetadata, error) {
-	googleapisDir := sources.Googleapis
-	if len(library.APIs) > 0 && library.APIs[0].Path == "schema/google/showcase/v1beta1" {
-		googleapisDir = sources.Showcase
-	}
-	metadata, err := repometadata.FromLibrary(cfg, library, googleapisDir)
-	if err != nil {
-		return nil, err
-	}
-
-	// Set fields not set by FromLibrary.
-	metadata.ClientDocumentation = fmt.Sprintf("https://docs.rs/%s/latest", library.Name)
-	metadata.LibraryType = repometadata.GAPICAutoLibraryType
-
-	return metadata, nil
 }
 
 // UpdateWorkspace updates dependencies for the entire Rust workspace.
